move client ctx to a package
